# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: ZC League API
  version: 1.0.0
  description: REST API for users, teams, players, tournaments, matches, results, stadiums, events, and admin management.
servers:
  - url: /api
    description: Default API base path
paths:
...
  /admins:
    get:
      summary: List admins
      responses:
        '200':
          description: Array of admins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Admin'
    post:
      summary: Create admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminCreate'
      responses:
        '201':
          description: Created admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
  /admins/{adminid}:
    get:
      summary: Get admin by id
      parameters:
        - $ref: '#/components/parameters/AdminId'
      responses:
        '200':
          description: Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '404':
          description: Not found
    patch:
      summary: Update admin
      parameters:
        - $ref: '#/components/parameters/AdminId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdate'
      responses:
        '200':
          description: Updated admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Admin'
        '404':
          description: Not found
    delete:
      summary: Delete admin
      parameters:
        - $ref: '#/components/parameters/AdminId'
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /teams:
    get:
      summary: List teams
      responses:
        '200':
          description: Array of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
    post:
      summary: Create team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
      responses:
        '201':
          description: Created team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
  /teams/{teamid}:
    get:
      summary: Get team by id
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '200':
          description: Team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          description: Not found
    patch:
      summary: Update team
      parameters:
        - $ref: '#/components/parameters/TeamId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamUpdate'
      responses:
        '200':
          description: Updated team
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          description: Not found
    delete:
      summary: Delete team
      parameters:
        - $ref: '#/components/parameters/TeamId'
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
...
  /players:
    get:
      summary: List players
      responses:
        '200':
          description: Array of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Player'
    post:
      summary: Create player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerCreate'
      responses:
        '201':
          description: Created player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
  /players/{playerid}:
    get:
      summary: Get player by id
      parameters:
        - $ref: '#/components/parameters/PlayerId'
      responses:
        '200':
          description: Player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          description: Not found
    patch:
      summary: Update player
      parameters:
        - $ref: '#/components/parameters/PlayerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlayerUpdate'
      responses:
        '200':
          description: Updated player
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Player'
        '404':
          description: Not found
    delete:
      summary: Delete player
      parameters:
        - $ref: '#/components/parameters/PlayerId'
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
...
  /matches:
    get:
      summary: List matches
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [Upcoming, Live, Finished]
        - in: query
          name: round
          schema:
            type: string
            enum: [Group, Quarter, Semi, Final]
      responses:
        '200':
          description: Array of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'
    post:
      summary: Create match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchCreate'
      responses:
        '201':
          description: Created match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
  /matches/{matchid}:
    get:
      summary: Get match by id
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '200':
          description: Match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '404':
          description: Not found
    patch:
      summary: Update match
      parameters:
        - $ref: '#/components/parameters/MatchId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchUpdate'
      responses:
        '200':
          description: Updated match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '404':
          description: Not found
    delete:
      summary: Delete match
      parameters:
        - $ref: '#/components/parameters/MatchId'
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /match-results:
    get:
      summary: List match results
      responses:
        '200':
          description: Array of match results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MatchResult'
    post:
      summary: Create match result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchResultCreate'
      responses:
        '201':
          description: Created match result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResult'
  /match-results/{resultid}:
    get:
      summary: Get match result by id
      parameters:
        - $ref: '#/components/parameters/ResultId'
      responses:
        '200':
          description: Match result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResult'
        '404':
          description: Not found
    patch:
      summary: Update match result
      parameters:
        - $ref: '#/components/parameters/ResultId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchResultUpdate'
      responses:
        '200':
          description: Updated match result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResult'
        '404':
          description: Not found
    delete:
      summary: Delete match result
      parameters:
        - $ref: '#/components/parameters/ResultId'
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /events:
    get:
      summary: List events
      responses:
        '200':
          description: Array of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
    post:
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Created event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
  /events/{eventid}:
    get:
      summary: Get event by id
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Not found
    patch:
      summary: Update event
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdate'
      responses:
        '200':
          description: Updated event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Event'
        '404':
          description: Not found
    delete:
      summary: Delete event
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
components:
  parameters:
...
    AdminId:
      in: path
      name: adminid
      required: true
      schema:
        type: integer
        format: int32
    TeamId:
      in: path
      name: teamid
      required: true
      schema:
        type: integer
        format: int32
    PlayerId:
      in: path
      name: playerid
      required: true
      schema:
        type: integer
        format: int32
...
    MatchId:
      in: path
      name: matchid
      required: true
      schema:
        type: integer
        format: int32
    ResultId:
      in: path
      name: resultid
      required: true
      schema:
        type: integer
        format: int32
    EventId:
      in: path
      name: eventid
      required: true
      schema:
        type: integer
        format: int32
  schemas:
...
    Admin:
      type: object
      properties:
        adminid: { type: integer, format: int32 }
        userid: { type: integer, format: int32 }
        permissionslevel: { type: integer, format: int32, default: 1 }
      required: [adminid]
    AdminCreate:
      type: object
      properties:
        userid: { type: integer, format: int32 }
        permissionslevel: { type: integer, format: int32 }
      required: [userid]
    AdminUpdate:
      type: object
      properties:
        permissionslevel: { type: integer, format: int32 }
    Team:
      type: object
      properties:
        teamid: { type: integer, format: int32 }
        teamname: { type: string }
        logourl: { type: string }
        teamcaptainid: { type: integer, format: int32, nullable: true }
        createdat: { type: string, format: date-time }
        description: { type: string }
      required: [teamid, teamname]
    TeamCreate:
      type: object
      properties:
        teamname: { type: string }
        logourl: { type: string }
        teamcaptainid: { type: integer, format: int32, nullable: true }
        description: { type: string }
      required: [teamname]
    TeamUpdate:
      type: object
      properties:
        teamname: { type: string }
        logourl: { type: string }
        teamcaptainid: { type: integer, format: int32, nullable: true }
        description: { type: string }
...
    Player:
      type: object
      properties:
        playerid: { type: integer, format: int32 }
        userid: { type: integer, format: int32 }
        teamid: { type: integer, format: int32, nullable: true }
        position: { type: string }
        jerseynumber: { type: integer, format: int32 }
        statsid: { type: integer, format: int32 }
        preferredfoot:
          type: string
          enum: [Left, Right, Both]
          default: Right
        height: { type: number, format: float }
        weight: { type: number, format: float }
      required: [playerid]
    PlayerCreate:
      type: object
      properties:
        userid: { type: integer, format: int32 }
        teamid: { type: integer, format: int32, nullable: true }
        position: { type: string }
        jerseynumber: { type: integer, format: int32 }
        statsid: { type: integer, format: int32 }
        preferredfoot: { type: string, enum: [Left, Right, Both] }
        height: { type: number, format: float }
        weight: { type: number, format: float }
      required: [userid, statsid]
    PlayerUpdate:
      type: object
      properties:
        teamid: { type: integer, format: int32, nullable: true }
        position: { type: string }
        jerseynumber: { type: integer, format: int32 }
        preferredfoot: { type: string, enum: [Left, Right, Both] }
        height: { type: number, format: float }
        weight: { type: number, format: float }
...
    Match:
      type: object
      properties:
        matchid: { type: integer, format: int32 }
        tournamentid: { type: integer, format: int32 }
        hometeamid: { type: integer, format: int32 }
        awayteamid: { type: integer, format: int32 }
        stadiumid: { type: integer, format: int32 }
        matchdate: { type: string, format: date-time }
        round: { type: string, enum: [Group, Quarter, Semi, Final], default: Group }
        status: { type: string, enum: [Upcoming, Live, Finished], default: Upcoming }
      required: [matchid, matchdate]
    MatchCreate:
      type: object
      properties:
        tournamentid: { type: integer, format: int32 }
        hometeamid: { type: integer, format: int32 }
        awayteamid: { type: integer, format: int32 }
        stadiumid: { type: integer, format: int32 }
        matchdate: { type: string, format: date-time }
        round: { type: string, enum: [Group, Quarter, Semi, Final] }
        status: { type: string, enum: [Upcoming, Live, Finished] }
      required: [matchdate]
    MatchUpdate:
      type: object
      properties:
        hometeamid: { type: integer, format: int32 }
        awayteamid: { type: integer, format: int32 }
        stadiumid: { type: integer, format: int32 }
        matchdate: { type: string, format: date-time }
        round: { type: string, enum: [Group, Quarter, Semi, Final] }
        status: { type: string, enum: [Upcoming, Live, Finished] }
    MatchResult:
      type: object
      properties:
        resultid: { type: integer, format: int32 }
        matchid: { type: integer, format: int32 }
        homescore: { type: integer, format: int32, default: 0 }
        awayscore: { type: integer, format: int32, default: 0 }
        winnerteamid: { type: integer, format: int32 }
        mvpplayerid: { type: integer, format: int32 }
      required: [resultid]
    MatchResultCreate:
      type: object
      properties:
        matchid: { type: integer, format: int32 }
        homescore: { type: integer, format: int32 }
        awayscore: { type: integer, format: int32 }
        winnerteamid: { type: integer, format: int32 }
        mvpplayerid: { type: integer, format: int32 }
      required: [matchid]
    MatchResultUpdate:
      type: object
      properties:
        homescore: { type: integer, format: int32 }
        awayscore: { type: integer, format: int32 }
        winnerteamid: { type: integer, format: int32 }
        mvpplayerid: { type: integer, format: int32 }